Split out read validation
